style: 코드 포맷팅, 파일 이름 변경
Close #171
