+++ src/main/java/com/teamtwo/model/ShippingAddressDAO.java
@@ -257,6 +257,13 @@ public class ShippingAddressDAO implements BaseDAO<ShippingAddressDTO> {
     }
   } // getupdate(id) end
 
+
+  /**
+   * 유저의 식별자를 인자로 받아서 해당하는 유저의 모든 배송지 리스트를 반환해주는 메소드입니다.
+   * 
+   * @param 유저의 식별자를 인자로 받습니다.
+   * @author vemm8ks2
+   */
   public List<ShippingAddressDTO> getShippingAddressByCustomer(int id) {
 
     List<ShippingAddressDTO> list = new ArrayList<>();
@@ -290,233 +297,5 @@ public class ShippingAddressDAO implements BaseDAO<ShippingAddressDTO> {
 
     return list;
   }
-
-		try {
-			open();
-
-			String sql = "SELECT * FROM T_shipping_address WHERE shipping_address_id = ?";
-
-			pstmt = conn.prepareStatement(sql);
-			pstmt.setInt(1, id);
-
-			rs = pstmt.executeQuery();
-
-			if (rs.next()) {
-				dto = new ShippingAddressDTO();
-
-				dto.setShippingAddressId(rs.getInt("shipping_address_id"));
-				dto.setShippingAddressDestination(rs.getString("shipping_address_destination"));
-				dto.setShippingAddressIsDefault(rs.getString("shipping_address_is_default"));
-				dto.setShippingAddressCustomerFk(rs.getInt("shipping_address_customerFk"));
-			}
-
-		} catch (Exception e) {
-			e.printStackTrace();
-		} finally {
-			close();
-		}
-
-		return dto;
-	}
-
-	/**
-	 * @author bborib
-	 */
-	@Override
-	public List<ShippingAddressDTO> getAll() {
-		List<ShippingAddressDTO> list = new ArrayList<>();
-
-		try {
-			open();
-
-			String sql = "SELECT * FROM T_shipping_address ORDER BY shipping_address_id DESC";
-
-			pstmt = conn.prepareStatement(sql);
-			rs = pstmt.executeQuery();
-
-			while (rs.next()) {
-				ShippingAddressDTO dto = new ShippingAddressDTO();
-
-				dto.setShippingAddressId(rs.getInt("shipping_address_id"));
-				dto.setShippingAddressDestination(rs.getString("shipping_address_destination"));
-				dto.setShippingAddressIsDefault(rs.getString("shipping_address_is_default"));
-				dto.setShippingAddressCustomerFk(rs.getInt("shipping_address_customerFk"));
-
-				list.add(dto);
-			}
-
-		} catch (Exception e) {
-			e.printStackTrace();
-		} finally {
-			close();
-		}
-
-		return list;
-	}
-
-	/**
-	 * @param 배송지 DTO를 인자로 받아서 새롭게 저장하는 메소드입니다.
-	 * @author bborib
-	 */
-	@Override
-	public void save(ShippingAddressDTO dto) {
-		try {
-			open();
-
-			String sql = "SELECT max(shipping_address_id) FROM T_shipping_address";
-
-			pstmt = conn.prepareStatement(sql);
-			rs = pstmt.executeQuery();
-
-			Integer count = null;
-
-			if (rs.next())
-				count = rs.getInt(1) + 1;
-
-			/**
-			 * TODO(24.09.03): 에러 핸들링
-			 */
-			if (count == null)
-				throw new SQLException();
-
-			sql = "INSERT INTO T_shipping_address VALUES(?, ?, ?, ?)";
-
-			pstmt = conn.prepareStatement(sql);
-
-			pstmt.setInt(1, count);
-			pstmt.setString(2, dto.getShippingAddressDestination());
-			pstmt.setString(3, dto.getShippingAddressIsDefault());
-			pstmt.setInt(4, dto.getShippingAddressCustomerFk());
-
-			pstmt.executeUpdate();
-
-		} catch (Exception e) {
-			e.printStackTrace();
-		} finally {
-			close();
-		}
-	}
-
-	/**
-	 * @param 배송지 DTO 객체를 인자로 받습니다.
-	 * @author bborib
-	 */
-	@Override
-	public void update(ShippingAddressDTO dto) {
-		try {
-			open();
-
-			/**
-			 * TODO(24.09.03): 유저의 외래키를 수정할 일은 없어야 합니다.
-			 */
-			String sql = "UPDATE T_shipping_address " + "SET shipping_address_destination = ?,"
-					+ "shipping_address_is_default = ?, " + "WHERE shipping_address_id = ?";
-
-			pstmt = conn.prepareStatement(sql);
-
-			pstmt.setString(1, dto.getShippingAddressDestination());
-			pstmt.setString(2, dto.getShippingAddressIsDefault());
-			pstmt.setInt(3, dto.getShippingAddressId());
-
-			pstmt.executeUpdate();
-
-		} catch (Exception e) {
-			e.printStackTrace();
-		} finally {
-			close();
-		}
-	}
-
-	/**
-	 * @param 배송지의 식별자를 인자로 받습니다.
-	 * @author bborib
-	 */
-	@Override
-	public void delete(int id) {
-		/**
-		 * TODO(24.09.03): 에러 발생 시 throw
-		 */
-
-		try {
-			open();
-
-			String sql = "DELETE FROM T_shipping_address WHERE shipping_address_id = ?";
-
-			pstmt = conn.prepareStatement(sql);
-
-			pstmt.setInt(1, id);
-			pstmt.executeUpdate();
-
-		} catch (Exception e) {
-			e.printStackTrace();
-		} finally {
-			close();
-		}
-	}
-
-	/**
-	 * 배송지의 식별자를 인자로 받아서 해당하는 배송지를 기본 배송지로 설정해주는 메소드입니다.
-	 * 
-	 * @param 배송지의 식별자를 인자로 받습니다.
-	 * @author joohp
-	 */
-	public void updateDefault(int id) {
-		/**
-		 * TODO(24.09.05): 배송지를 새롭게 'Y'로 업데이트 해줄 뿐만 아니라 기존에 있던 기본 배송지도 'N'으로 업데이트해야 한다.
-		 */
-
-		try {
-			open();
-
-			String sql = "UPDATE T_shipping_address SET shipping_address_is_default = ? WHERE shipping_address_id = ?";
-
-			pstmt = conn.prepareStatement(sql);
-
-			pstmt.setString(1, "Y");
-			pstmt.setInt(2, id);
-
-			pstmt.executeUpdate();
-
-		} catch (SQLException e) {
-			e.printStackTrace();
-		} finally {
-			close();
-		}
-	} // getupdate(id) end
-
-	public List<ShippingAddressDTO> getcustomeraddr(int fk) {
-		List<ShippingAddressDTO> list = new ArrayList<>();
-		try {
-			open();
-			String sql = "SELECT * FROM T_shipping_address WHERE shipping_address_customer_fk = ?";
-
-			pstmt = conn.prepareStatement(sql);
-			pstmt.setInt(1, fk);
-			rs = pstmt.executeQuery();
-
-			while (rs.next()) {
-				ShippingAddressDTO dto = new ShippingAddressDTO();
-				dto.setShippingAddressId(rs.getInt("shipping_address_id"));
-				dto.setShippingAddressDestination(rs.getString("shipping_address_destination"));
-				dto.setShippingAddressIsDefault(rs.getString("shipping_address_is_default"));
-				dto.setShippingAddressCustomerFk(rs.getInt("shipping_address_customerFk"));
-
-				list.add(dto);
-			}
-		} catch (Exception e) {
-			e.printStackTrace();
-		} finally {
-			close();
-		}
-
-		return list;
-	}
-	
-	
-	
-	
-	
-	
-	
 
 }
